Add HTML stripping as an option for analyzer settings Add default standard analyzer with html stripping built in Refactor ZuliaPerFieldAnalyzer to split out the tokenization chain into ZuliaFieldAnalyzer Make sure in analysis requests that a custom analyzer used would be closed
